lambda-workshop/layer_requirements.txt
requests
boto3>=1.26.129
aws-psycopg2>=1.3.8